Introduced fastnetmon_logic which will keep keep all business logic functions
